clean index.js
